+++ ui/src/App.js
@@ -158,12 +158,14 @@ function App() {
 
       // Case 1: we need authorization for exporting the recovery phrase.
       if (event.payload === GET_RECOVERY_PHRASE) {
+        console.log("[INFO]: Viewing secret phrase window");
         navigate("/view-secret-phrase");
         appWindow.show();
         return;
       }
 
       // Case 2: we need authorization for signing a transaction.
+      console.log("[INFO]: Authorization for transaction");
       await checkAndStopExportingPhrase();
       let parsedAuthorizationSummary = parseTransactionSummary(event.payload.split(" "));
 
@@ -175,7 +177,7 @@ function App() {
 
   const listenForShowSecretPhraseRequests = async () => {
     console.log("[INFO]: Setup tray show secret phrase listener.");
-    listen('show_secret_phrase', (_event) => {
+    listen('show_secret_phrase', async (_event) => {
       getSecretRecoveryPhrase();
     })
   }
@@ -207,11 +209,17 @@ function App() {
       exportingPhraseRef.current = true;
     }
 
-    console.log("[INFO]: Send request to export recovery phrase.");
-    let phrase = await invoke('get_recovery_phrase', { prompt: GET_RECOVERY_PHRASE });
+    try {
+      console.log("[INFO]: Send request to export recovery phrase.");
+      let phrase = await invoke('get_recovery_phrase', { prompt: GET_RECOVERY_PHRASE })
 
-    if (phrase) {
-      setExportedSecretPhrase(phrase);
+      console.log("[INFO]: Request for recovery phrase passed");
+      if (phrase) {
+        await setExportedSecretPhrase(phrase);
+      }
+    } catch(error) {
+      console.log("[WARNING]: Failed getting mnemonic, either fail or aborted (expected)");
+      return;
     }
   }